Sorted the scaling of the dodecahedron, and fixed bounding volume
Is was radius = sqrt(3) by default, so added a scaling factor.
Also, closed the bounding sphere tight to the obecjt - it was using the naive bounding
algorithm in the SPc.Combine function, which constructs a cuboid around the object, then
adds a sphere.
Hmmm. Actually that gives me an idea of how to improve that function...